revert data type changing

+++ test/CyberEngine.t.sol
@@ -173,7 +173,7 @@ contract CyberEngineTest is Test {
 
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob_handle";
         bytes32 digest = engine.hashTypedDataV4(
@@ -189,7 +189,7 @@ contract CyberEngineTest is Test {
     function testCannotVerifyAsNonSigner() public {
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob_handle";
         bytes32 digest = engine.hashTypedDataV4(
@@ -207,7 +207,7 @@ contract CyberEngineTest is Test {
     function testCannotVerifyDeadlinePassed() public {
         // change block timestamp to make deadline invalid
         vm.warp(150);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob_handle";
         bytes32 digest = engine.hashTypedDataV4(
@@ -337,7 +337,7 @@ contract CyberEngineTest is Test {
 
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob";
         bytes32 digest = engine.hashTypedDataV4(
@@ -369,7 +369,7 @@ contract CyberEngineTest is Test {
 
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob_handle";
         bytes32 digest = engine.hashTypedDataV4(
@@ -395,7 +395,7 @@ contract CyberEngineTest is Test {
 
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob_handle";
         bytes32 digest = engine.hashTypedDataV4(
@@ -425,7 +425,7 @@ contract CyberEngineTest is Test {
 
         // change block timestamp to make deadline valid
         vm.warp(50);
-        uint64 deadline = 100;
+        uint256 deadline = 100;
 
         string memory handle = "bob";
         bytes32 digest = engine.hashTypedDataV4(
